Move parse_model into utils
